iOS: improve backend robustness, logging, and Metal safety
- Replace iOS NSLog traces with mfb_log.
- Add safer buffer/layout validation and allocation checks.
- Add iOS timer state to window data and improve sync/wait handling.
- Harden resize flow: handle texture resize failures and keep state consistent.
- Add ARC guards in resize/dealloc texture paths.
- Improve multi-touch handling with stable touch-to-button mapping and edge-case guards.
- Defer resize callback from render thread to update/event path.
- Minor build/internal cleanups (skip examples on iOS, explicit interpolation casts).

+++ src/MiniFB_internal.c
@@ -19,10 +19,10 @@ interpolate(uint32_t *src_image, uint32_t x, uint32_t y, uint32_t src_offset_x, 
     uint32_t wx1 = 0x10000 - wx2;
     uint32_t wy1 = 0x10000 - wy2;
 
-    uint32_t w1 = ((uint64_t) wx1 * wy1) >> 16;
-    uint32_t w2 = ((uint64_t) wx2 * wy1) >> 16;
-    uint32_t w3 = ((uint64_t) wx1 * wy2) >> 16;
-    uint32_t w4 = ((uint64_t) wx2 * wy2) >> 16;
+    uint32_t w1 = (uint32_t)(((uint64_t) wx1 * wy1) >> 16);
+    uint32_t w2 = (uint32_t)(((uint64_t) wx2 * wy1) >> 16);
+    uint32_t w3 = (uint32_t)(((uint64_t) wx1 * wy2) >> 16);
+    uint32_t w4 = (uint32_t)(((uint64_t) wx2 * wy2) >> 16);
 
     // If you don't have uint64_t
     //uint32_t b = (((p00[0] * wx1 + p01[0] * wx2) >> 16) * wy1 + ((p10[0] * wx1 + p11[0] * wx2) >> 16) * wy2) >> 16;
